implemented get family nutritions average

+++ src/main/java/it/da/be/challenge/fruit/dto/NutritionsInfoProjection.java
@@ -5,7 +5,7 @@ public interface NutritionsInfoProjection {
 	double getCarbohydrates();
     double getProtein();
     double getFat();
-    int getCalories();
+    double getCalories();
     double getSugar();
 
 }
